Year-1/Semester-1/Information-Retrieval/fmiunisofia-mini-course-project/src/config.py
import os
from dotenv import load_dotenv

load_dotenv()

ES_URL = os.getenv("ES_URL", "http://localhost:9200")
ES_USER = os.getenv("ES_USER") or None
ES_PASS = os.getenv("ES_PASS") or None
ES_INDEX = os.getenv("ES_INDEX", "semeval_task10_docs")

