Year-1/Semester-1/Information-Retrieval/fmiunisofia-mini-course-project/requirements.txt
elasticsearch>=8.0.0,<9.0.0
python-dotenv>=1.0.0
flask>=3.0.0
tqdm>=4.66.0

